Refactor native PHP child helpers
